models/memory/agents/agents/agents/agents/requirements.txt
New-Item -Path "requirements.txt" -ItemType "file" -Force
@"
langchain==1.0.0
langchain-core==0.3.0
langchain-groq==0.6.0
pydantic==2.5.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
"@ | Out-File -FilePath requirements.txt -Encoding utf8